Move MockStsCredentialsFetcher.java
